Rework EuclideanRing laws, fixes #99.
This change that fixes the EuclideanRing laws so that they do actually
describe a euclidean ring (I have discussed some problems with the
current set of laws in #99).

This change to the laws actually invalidates the `EuclideanRing Unit`
and `Field Unit` instances, but they remain in for now so that we can
publish this without it being a breaking change. These instances have
been marked as deprecated and will be removed in the next major release.

Also add more explanation about division by zero in the EuclideanRing
class.