Improve size calculation for compacting
When we are measuring the size of a subtree in order to decide whether
to compact it after pretty-printing, we used to double the size each
time we move up a level. This is, in retrospect, not ideal, because
labels such as AssocProp or the Delta labels like ExtraLeft or Same
don't actually represent any additional depth in the tree. Now we just
sum up all of the nodes in a subtree.